Explain statistical output

Use an LLM to explain the output from various statistical objects using straightforward, understandable, and context-aware natural language descriptions.

Usage

explain(object, chat, ...)

# S3 method for class 'htest'
explain(object, chat, context = NULL, echo = NULL, verbose = FALSE, ...)

# S3 method for class 'lm'
explain(object, chat, context = NULL, echo = NULL, verbose = FALSE, ...)

# S3 method for class 'glm'
explain(object, chat, context = NULL, echo = NULL, verbose = FALSE, ...)

# S3 method for class 'polr'
explain(object, chat, context = NULL, echo = NULL, verbose = FALSE, ...)

# S3 method for class 'lme'
explain(object, chat, context = NULL, echo = NULL, verbose = FALSE, ...)

# S3 method for class 'lmerMod'
explain(object, chat, context = NULL, echo = NULL, verbose = FALSE, ...)

# S3 method for class 'glmerMod'
explain(object, chat, context = NULL, echo = NULL, verbose = FALSE, ...)

# S3 method for class 'gam'
explain(object, chat, context = NULL, echo = NULL, verbose = FALSE, ...)

# S3 method for class 'survreg'
explain(object, chat, context = NULL, echo = NULL, verbose = FALSE, ...)

Arguments

object

An appropriate statistical object. For example, object can be the output from calling t.test() or glm().

chat

A Chat object (e.g., from calling chat_openai() or [chat_gemini()][ellmer::chat_gemini)]).

[ellmer::chat_gemini)]: R:ellmer::chat_gemini)

...

Additional optional arguments. (Currently ignored.)

context

Optional character string providing additional context, such as background on the experiment and information about the data.

echo

Logical specifying whether to emit the response to stdout as it's received. If NULL (the default), then the value of echo set when the chat object was created will be used. See the $chat() method described in Chat.

verbose

Logical indicating whether to print the system prompt and chatbot input, which can be useful for debugging. Default is FALSE.

Examples

if (FALSE) { # \dontrun{
# Polynomial regression
cars_lm <- lm(dist ~ poly(speed, degree = 2), data = cars)
context <- "
The data give the speed of cars (mph) and the distances taken to stop (ft).
Note that the data were recorded in the 1920s.
"
# Use Google Gemini to explain the output; requires an API key; see
# ?ellmer::chat_gemini for details
chat <- ellmer::chat_gemini()
explain(cars_lm, chat = chat, context = context)

# Poisson regression example from ?stats::glm
counts <- c(18,17,15,20,10,20,25,13,12)
outcome <- gl(3,1,9)
treatment <- gl(3,3)
data.frame(treatment, outcome, counts) # showing data
D93_glm <- glm(counts ~ outcome + treatment, family = poisson())

# Use Google Gemini to explain the output; requires an API key; see
# ?ellmer::chat_gemini for details
chat <- ellmer::chat_gemini()
explain(D93_glm, chat = chat, verbose = TRUE)
} # }